requirements_dev.txt
django-debug-toolbar==4.4.6
Faker==30.3.0
coverage==7.6.2
